Modified UI a little bit
* Question
* Timer

+++ app/src/main/java/com/carlosjordi/monthlychallenge03/ui/screens/quiz/QuizState.kt
@@ -3,7 +3,7 @@ package com.carlosjordi.monthlychallenge03.ui.screens.quiz
 import com.carlosjordi.monthlychallenge03.data.model.Question
 
 data class QuizState(
-    val timer: Int = 30,
+    val timer: Float = 30f,
     val isOptionSelected: Boolean = false,
     val isRightAnswer: Boolean? = null,
     val currentQuestion: Question,
